+++ inst/testscripts/sourcescript1.R
@@ -1,1 +1,1 @@
-source("source1.r")
+source("source2.r")