+++ NEWS.md
@@ -1,3 +1,8 @@
+# provParseR 0.2
+
+* Added functions: get.stdout.nodes, get.args, get.val.type, get.tool.info
+* Snapshot nodes include path to the snapshot file
+
 # provParseR 0.1.2
 
 * Added functions: get.urls, get.error.nodes, and get.variable.named